- More NDK/MSDDK compatibility fixes.
svn path=/trunk/; revision=17652

+++ reactos/include/ndk/rtlfuncs.h
@@ -93,7 +93,7 @@ DWORD
 STDCALL
 RtlCompactHeap(
     HANDLE heap,
-    DWORD flags
+    ULONG flags
 );
 
 HANDLE
